Move theme example to top level theme switch
