+++ examples/api/src/views/App.svelte
@@ -1,10 +1,8 @@
 <script lang="ts">
-  import { show, hide, setTheme, setDockVisibility } from '@tauri-apps/api/app'
-  import type { Theme } from '@tauri-apps/api/window'
+  import { show, hide, setDockVisibility } from '@tauri-apps/api/app'
   import type { ViewProps } from '../types'
 
   let { onMessage }: ViewProps = $props()
-  let theme = $state<Theme | 'auto'>('auto')
   let dockVisible = $state(true)
 
   function showApp() {
@@ -25,21 +23,6 @@
       .catch(onMessage)
   }
 
-  async function switchTheme() {
-    switch (theme) {
-      case 'dark':
-        theme = 'light'
-        break
-      case 'light':
-        theme = 'auto'
-        break
-      case 'auto':
-        theme = 'dark'
-        break
-    }
-    setTheme(theme === 'auto' ? null : theme)
-  }
-
   async function toggleDockVisibility() {
     await setDockVisibility(!dockVisible)
     dockVisible = !dockVisible
@@ -54,9 +37,6 @@
     onclick={showApp}>Show</button
   >
   <button class="btn" id="hide" onclick={hideApp}>Hide</button>
-  <button class="btn" id="switch-theme" onclick={switchTheme}
-    >Switch Theme ({theme})</button
-  >
   <button class="btn" id="toggle-dock-visibility" onclick={toggleDockVisibility}
     >Toggle dock visibility</button
   >
